+++ src/QueryBuilderServiceProvider.php
@@ -72,5 +72,21 @@ class QueryBuilderServiceProvider extends ServiceProvider
 
             return collect($columns);
         });
+      
+        Request::macro('sorts', function ($default = null) {
+            $sortParts = $this->sort();
+
+            if (! is_array($sortParts)) {
+                $sortParts = explode(',', $sortParts);
+            }
+
+            $sorts = collect($sortParts)->filter();
+
+            if ($sorts->isNotEmpty()) {
+                return $sorts;
+            }
+
+            return collect($default)->filter();
+        });
     }
 }