+++ tests/FilterTest.php
@@ -7,7 +7,7 @@ use Spatie\QueryBuilder\Filter;
 use Spatie\QueryBuilder\QueryBuilder;
 use Illuminate\Database\Eloquent\Builder;
 use Spatie\QueryBuilder\Tests\Models\TestModel;
-use Spatie\QueryBuilder\Exceptions\InvalidQuery;
+use Spatie\QueryBuilder\Exceptions\InvalidFilterQuery;
 use Spatie\QueryBuilder\Filters\Filter as FilterInterface;
 
 class FilterTest extends TestCase
@@ -148,13 +148,74 @@ class FilterTest extends TestCase
     }
 
     /** @test */
+    public function it_can_allow_multiple_filters()
+    {
+        $model1 = TestModel::create(['name' => 'abcdef']);
+        $model2 = TestModel::create(['name' => 'abcdef']);
+
+        $results = $this
+            ->createQueryFromFilterRequest([
+                'name' => 'abc',
+            ])
+            ->allowedFilters('name', Filter::exact('id'))
+            ->get();
+
+        $this->assertCount(2, $results);
+        $this->assertEquals([$model1->id, $model2->id], $results->pluck('id')->all());
+    }
+
+    /** @test */
+    public function it_can_allow_multiple_filters_as_an_array()
+    {
+        $model1 = TestModel::create(['name' => 'abcdef']);
+        $model2 = TestModel::create(['name' => 'abcdef']);
+
+        $results = $this
+            ->createQueryFromFilterRequest([
+                'name' => 'abc',
+            ])
+            ->allowedFilters(['name', Filter::exact('id')])
+            ->get();
+
+        $this->assertCount(2, $results);
+        $this->assertEquals([$model1->id, $model2->id], $results->pluck('id')->all());
+    }
+
+    /** @test */
+    public function it_can_filter_by_multiple_filters()
+    {
+        $model1 = TestModel::create(['name' => 'abcdef']);
+        $model2 = TestModel::create(['name' => 'abcdef']);
+
+        $results = $this
+            ->createQueryFromFilterRequest([
+                'name' => 'abc',
+                'id' => "1,{$model1->id}",
+            ])
+            ->allowedFilters('name', Filter::exact('id'))
+            ->get();
+
+        $this->assertCount(1, $results);
+        $this->assertEquals([$model1->id], $results->pluck('id')->all());
+    }
+
+    /** @test */
     public function it_guards_against_invalid_filters()
     {
-        $this->expectException(InvalidQuery::class);
+        $this->expectException(InvalidFilterQuery::class);
 
         $this
             ->createQueryFromFilterRequest(['name' => 'John'])
             ->allowedFilters('id');
+    }
+
+    /** @test */
+    public function an_invalid_filter_query_exception_contains_the_unknown_and_allowed_filters()
+    {
+        $exception = new InvalidFilterQuery(collect(['unknown filter']), collect(['allowed filter']));
+
+        $this->assertEquals(['unknown filter'], $exception->unknownFilters->all());
+        $this->assertEquals(['allowed filter'], $exception->allowedFilters->all());
     }
 
     protected function createQueryFromFilterRequest(array $filters): QueryBuilder