+++ tests/Models/RelatedModel.php
@@ -3,6 +3,8 @@
 namespace Spatie\QueryBuilder\Tests\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class RelatedModel extends Model
 {
@@ -13,5 +15,10 @@ class RelatedModel extends Model
     public function testModel(): BelongsTo
     {
         return $this->belongsTo(TestModel::class);
+    }
+
+    public function nestedRelatedModels(): HasMany
+    {
+        return $this->hasMany(NestedRelatedModel::class);
     }
 }